contractor datacontext

+++ DFAPI/Models/Master.cs
@@ -138,7 +138,7 @@ namespace DFAPI.Entities
     {
         [Key]
         public long ID { get; set; }
-        public int? UserType { get; set; }
+        public byte? UserType { get; set; }
         public long? UserId { get; set; }
         public int? DepartmentID { get; set; }
         public bool? Display { get; set; }
@@ -166,7 +166,7 @@ namespace DFAPI.Entities
     {
         [Key]
         public long ID { get; set; }
-        public int? UserType { get; set; }
+        public byte? UserType { get; set; }
         public long? UserId { get; set; }
         public int? DesignationID { get; set; }
         public bool? ReportingAuthority { get; set; }
